isolatePhase4: new File for HOC
