+++ front/src/applications/osrd/components/Simulation/SignalSwitch/SignalSwitch.js
@@ -1,33 +1,7 @@
 import { LIST_VALUES_SIGNAL_BASE, SIGNAL_BASE_DEFAULT } from 'applications/osrd/components/Simulation/consts';
 import PropTypes, { string } from 'prop-types';
-import { useDispatch, useSelector } from 'react-redux';
 
 import SwitchSNCF from 'common/BootstrapSNCF/SwitchSNCF/SwitchSNCF'
-import { updateSignalBase } from "reducers/osrdsimulation";
-
-/**
- * HOC to provide store data
- * @param {RFC} Component
- * @returns RFC with OSRD Data. SignalSwitch
- */
-const withOSRDData = (Component) => (props) => {
-  const {signalBase} = useSelector((state) => state.osrdsimulation);
-  const dispatch = useDispatch();
-  /**
-   * Store update on toggle
-   * @param {SyntheticBaseEvent} e the Event triggered by the signal UI
-   */
-   const toggleSignal = (e) => {
-    const newSignal = e?.target?.value;
-    if (typeof newSignal !== 'undefined') {
-      dispatch(updateSignalBase(newSignal));
-    }
-    else {
-      console.warn('Try to toggle Signal with unavailableValue')
-    }
-  };
-  return <Component {...props} signalBase={signalBase} toggleSignal={toggleSignal}></Component>
-}
 
 /**
  * RFC to display a switch adapted to Signal Display Logic
@@ -66,6 +40,5 @@ SignalSwitch.defaultProps = {
   type: "inline"
 }
 
-const OSRDSwitchSNCF = withOSRDData(SignalSwitch)
+export default SignalSwitch
 
-export default OSRDSwitchSNCF
